there is no such thing as cdist install in master
remove all traces of it to prevent the ongoing merge issues when using the real thing
