+++ cdist/core/cdist_type.py
@@ -100,11 +100,6 @@ class CdistType(object):
         return os.path.isfile(os.path.join(self.absolute_path, "singleton"))
 
     @property
-    def is_install(self):
-        """Check whether a type is used for installation (if not: for configuration)"""
-        return os.path.isfile(os.path.join(self.absolute_path, "install"))
-
-    @property
     def explorers(self):
         """Return a list of available explorers"""
         if not self.__explorers:
