+++ cdist/test/cdist_type/__init__.py
@@ -106,16 +106,6 @@ class TypeTestCase(test.CdistTestCase):
         cdist_type = core.CdistType(base_path, '__not_singleton')
         self.assertFalse(cdist_type.is_singleton)
 
-    def test_install_is_install(self):
-        base_path = fixtures
-        cdist_type = core.CdistType(base_path, '__install')
-        self.assertTrue(cdist_type.is_install)
-
-    def test_not_install_is_install(self):
-        base_path = fixtures
-        cdist_type = core.CdistType(base_path, '__not_install')
-        self.assertFalse(cdist_type.is_install)
-
     def test_with_explorers(self):
         base_path = fixtures
         cdist_type = core.CdistType(base_path, '__with_explorers')
