+++ cdist/test/config/fixtures/object/__third/moon/.cdist/parameter/name
