+++ scripts/cdist
@@ -26,7 +26,6 @@ def commandline():
 
     import cdist.banner
     import cdist.config
-    # import cdist.install
     import cdist.shell
 
     # Construct parser others can reuse
@@ -53,39 +52,35 @@ def commandline():
         parents=[parser['loglevel']])
     parser['banner'].set_defaults(func=cdist.banner.banner)
 
-    # Config and install (common stuff)
-    parser['configinstall'] = argparse.ArgumentParser(add_help=False)
-    parser['configinstall'].add_argument('host', nargs='+',
+    # Config
+    parser['config'] = parser['sub'].add_parser('config',
+        parents=[parser['loglevel']])
+    parser['config'].add_argument('host', nargs='+',
         help='one or more hosts to operate on')
-    parser['configinstall'].add_argument('-c', '--conf-dir',
+    parser['config'].add_argument('-c', '--conf-dir',
          help='Add configuration directory (can be repeated, last one wins)',
          action='append')
-    parser['configinstall'].add_argument('-i', '--initial-manifest', 
+    parser['config'].add_argument('-i', '--initial-manifest', 
          help='Path to a cdist manifest or \'-\' to read from stdin.',
          dest='manifest', required=False)
-    parser['configinstall'].add_argument('-n', '--dry-run',
+    parser['config'].add_argument('-n', '--dry-run',
          help='Do not execute code', action='store_true')
-    parser['configinstall'].add_argument('-o', '--out-dir',
+    parser['config'].add_argument('-o', '--out-dir',
          help='Directory to save cdist output in', dest="out_path")
-    parser['configinstall'].add_argument('-p', '--parallel',
+    parser['config'].add_argument('-p', '--parallel',
          help='Operate on multiple hosts in parallel',
          action='store_true', dest='parallel')
-    parser['configinstall'].add_argument('-s', '--sequential',
+    parser['config'].add_argument('-s', '--sequential',
          help='Operate on multiple hosts sequentially (default)',
          action='store_false', dest='parallel')
-
-    parser['configinstall'].add_argument('--remote-copy',
+    parser['config'].add_argument('--remote-copy',
          help='Command to use for remote copy (should behave like scp)',
          action='store', dest='remote_copy',
          default=cdist.REMOTE_COPY)
-    parser['configinstall'].add_argument('--remote-exec',
+    parser['config'].add_argument('--remote-exec',
          help='Command to use for remote execution (should behave like ssh)',
          action='store', dest='remote_exec',
          default=cdist.REMOTE_EXEC)
-
-    # Config
-    parser['config'] = parser['sub'].add_parser('config',
-        parents=[parser['loglevel'], parser['configinstall']])
     parser['config'].set_defaults(func=cdist.config.Config.commandline)
 
     # Shell
@@ -95,12 +90,6 @@ def commandline():
          help='Select shell to use, defaults to current shell')
     parser['shell'].set_defaults(func=cdist.shell.Shell.commandline)
 
-
-    # Install
-    # 20120525/sar: commented until it actually does something
-    #parser['install'] = parser['sub'].add_parser('install',
-    #    parents=[parser['loglevel'], parser['configinstall']])
-    #parser['install'].set_defaults(func=install)
 
     for p in parser:
         parser[p].epilog = "Get cdist at http://www.nico.schottelius.org/software/cdist/"
@@ -132,9 +121,6 @@ def commandline():
         sys.exit(0)
 
     args.func(args)
-
-#def install(args):
-#    configinstall(args, mode=cdist.install.Install)
 
 def emulator():
     """Prepare and run emulator"""
